--- app/src/main/java/com/example/bundlebundle/ui/home/ProductGridFragment.java
@@ -1,6 +1,0 @@
-package com.example.bundlebundle.ui.home;
-
-import androidx.fragment.app.Fragment;
-
-public class ProductGridFragment extends Fragment {
-}
