+++ app/src/main/java/com/example/bundlebundle/ui/sample/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.bundlebundle.ui.gallery
+package com.example.bundlebundle.ui.sample
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
