--- app/src/main/java/com/example/bundlebundle/ui/slideshow/SlideshowViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.bundlebundle.ui.slideshow
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class SlideshowViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
-    }
-    val text: LiveData<String> = _text
-}
