+++ app/src/main/res/layout/fragment_product_grid_list.xml
@@ -6,8 +6,6 @@
     android:name="com.example.bundlebundle.ProductGridFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:layout_marginLeft="16dp"
-    android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".ProductGridFragment"
+    tools:context=".ui.product.ProductGridFragment"
     tools:listitem="@layout/fragment_product_grid" />
